Separate folder for types
