+++ src/components/favorite-card.tsx
@@ -1,4 +1,4 @@
-import { OfferData } from '../types';
+import { OfferData } from '../types/types';
 import { Link } from 'react-router-dom';
 
 type FavCardProps = {
